refactor: organize `utils` namespace
