+++ README.md
@@ -237,7 +237,7 @@ You may pass your implementation of `Fetcher` function to option `fetcher`.
 ```ts
 import { GoogleTranslator } from 'anylang/translators';
 
-import { Fetcher, basicFetcher } from 'anylang/utils/fetcher';
+import { Fetcher, basicFetcher } from 'anylang/utils';
 
 // Extend `basicFetcher` with use CORS proxy for all requests
 const fetcher: Fetcher = async (url, options) => {
@@ -390,7 +390,7 @@ This class is just a decorator over `Scheduler` that allow you to use standard o
 import { GoogleTranslator } from 'anylang/translators';
 import { Scheduler } from 'anylang/scheduling/Scheduler';
 import { SchedulerWithCache } from 'anylang/scheduling/SchedulerWithCache';
-import { ICache } from 'anylang/utils/Cache';
+import { ICache } from 'anylang/utils';
 
 // Dummy implementation of cache
 class CacheInRam implements ICache {
