+++ src/utils/fetcher/index.ts
@@ -1,0 +1,2 @@
+export * from './basicFetcher';
+export * from './types';
